Assign default randomized imageUrl in create function and clean up unused code.
